% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression Function}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula object describing the model.}

\item{data}{A data frame containing the variables in the model.}
}
\value{
An object of class "linreg" with computed regression statistics.
}
\description{
This function performs linear regression using ordinary least squares.
}
\examples{
data(mtcars)
fit <- linreg(mpg ~ wt + hp, data = mtcars)
print(fit)
}
