% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Print method for linreg objects}
\usage{
\method{print}{linreg}(x)
}
\arguments{
\item{x}{An object of class "linreg".}
}
\description{
Prints the coefficients of the model.
}
